Moved Timestep tests into new structure
Coverage shouldn't change!

Issue #466
